user management change

+++ src/views/UserManagement.vue
@@ -18,7 +18,22 @@
         </template>
       </el-input>
 
-      <el-button type="primary" :icon="Plus" @click="showAddDialog">+ New</el-button>
+      <div style="display: flex; gap: 10px;">
+        <!-- Refresh Button -->
+        <el-tooltip content="Refresh User table" placement="top">
+          <el-button
+              class="refresh-button"
+              type="primary"
+              circle
+              @click="fetchUserData"
+          >
+            <el-icon style="color: #004085;"><RefreshRight /></el-icon>
+          </el-button>
+        </el-tooltip>
+
+        <!-- Add Button -->
+        <el-button type="primary" :icon="Plus" @click="showAddDialog">+ New</el-button>
+      </div>
     </div>
 
     <!-- Table -->
@@ -111,21 +126,21 @@
     </el-table>
 
     <!-- Add User Dialog -->
-    <el-dialog title="Add User" v-model="addDialogVisible" width="50%">
+    <el-dialog title="Add User" v-model="addDialogVisible" width="50%" @keyup.enter.native="validateAndAddUser">
       <div class="popup-container">
-        <el-form :model="newUser" label-width="140px">
-          <el-form-item label="Name">
+        <el-form ref="addUserForm" :model="newUser" :rules="rules" label-width="140px">
+          <el-form-item label="Name" prop="name">
             <el-input v-model="newUser.name" />
           </el-form-item>
 
-          <el-form-item label="Role">
+          <el-form-item label="Role" prop="role">
             <el-select v-model="newUser.role" placeholder="Select Role">
               <el-option label="管理员" value="管理员" />
               <el-option label="质检人员" value="质检人员" />
             </el-select>
           </el-form-item>
 
-          <el-form-item label="WeChat ID">
+          <el-form-item label="WeChat ID" prop="wecomId">
             <el-input v-model="newUser.wecomId" />
           </el-form-item>
 
@@ -137,18 +152,18 @@
             <el-input v-model="newUser.phone_number" />
           </el-form-item>
 
-          <el-form-item label="Status">
+          <el-form-item label="Status" prop="status">
             <el-select v-model="newUser.status" placeholder="Select Status">
               <el-option label="Active" :value="1" />
               <el-option label="Inactive" :value="0" />
             </el-select>
           </el-form-item>
 
-          <el-form-item label="Username">
+          <el-form-item label="Username" prop="username">
             <el-input v-model="newUser.username" />
           </el-form-item>
 
-          <el-form-item label="Password">
+          <el-form-item label="Password" prop="password">
             <el-input v-model="newUser.password" type="password" show-password />
           </el-form-item>
         </el-form>
@@ -157,31 +172,31 @@
       <template #footer>
         <div class="popup-container">
           <el-button @click="addDialogVisible = false">Cancel</el-button>
-          <el-button type="primary" @click="handleAdd">Confirm</el-button>
+          <el-button type="primary" @click="validateAndAddUser">Confirm</el-button>
         </div>
       </template>
     </el-dialog>
 
     <!-- Edit User Dialog -->
-    <el-dialog title="Edit User" v-model="editDialogVisible" width="50%">
+    <el-dialog title="Edit User" v-model="editDialogVisible" width="50%" @keyup.enter.native="handleEditConfirm">
       <div class="popup-container">
-        <el-form :model="editUser" label-width="140px">
-          <el-form-item label="Name">
+        <el-form ref="editUserForm" :model="editUser" :rules="rules" label-width="140px">
+          <el-form-item label="Name" prop="name">
             <el-input v-model="editUser.name" />
           </el-form-item>
 
-          <el-form-item label="Role">
+          <el-form-item label="Role" prop="role">
             <el-select v-model="editUser.role" placeholder="Select Role">
               <el-option label="管理员" value="管理员" />
               <el-option label="质检人员" value="质检人员" />
             </el-select>
           </el-form-item>
 
-          <el-form-item label="WeChat ID">
+          <el-form-item label="WeCom ID" prop="wecomId">
             <el-input v-model="editUser.wecomId" />
           </el-form-item>
 
-          <el-form-item label="Username">
+          <el-form-item label="Username" prop="username">
             <el-input v-model="editUser.username" />
           </el-form-item>
 
@@ -193,7 +208,7 @@
             <el-input v-model="editUser.phone_number" />
           </el-form-item>
 
-          <el-form-item label="Status">
+          <el-form-item label="Status" prop="status">
             <el-select v-model="editUser.status" placeholder="Select Status">
               <el-option label="Active" :value="1" />
               <el-option label="Inactive" :value="0" />
@@ -206,11 +221,11 @@
           </el-form-item>
 
           <!-- Password Fields -->
-          <el-form-item v-if="changePassword" label="New Password">
+          <el-form-item v-if="changePassword" label="New Password" prop="newPassword">
             <el-input v-model="newPassword" type="password" show-password />
           </el-form-item>
 
-          <el-form-item v-if="changePassword" label="Confirm Password">
+          <el-form-item v-if="changePassword" label="Confirm Password" prop="confirmPassword">
             <el-input v-model="confirmPassword" type="password" show-password />
           </el-form-item>
 
@@ -229,13 +244,14 @@
 </template>
 
 <script>
-import { Search, Plus, QuestionFilled } from '@element-plus/icons-vue';
+import { Search, Plus, QuestionFilled, RefreshRight } from '@element-plus/icons-vue';
 import {
   fetchUsers,
   addUser,
   updateUser,
   deleteUser,
 } from '@/services/userService.js';
+import { ElMessage, ElMessageBox } from "element-plus";
 
 export default {
   name: 'UserManagement',
@@ -243,6 +259,7 @@ export default {
     QuestionFilled,
     Search,
     Plus,
+    RefreshRight
   },
   data() {
     return {
@@ -269,11 +286,51 @@ export default {
         username: '',
         email: '',
         phone_number: '',
-        status: null,
+        status: null
       },
       changePassword: false, // Checkbox state for edit dialog
       newPassword: '', // New password input
       confirmPassword: '', // Confirm password input
+      rules: {
+        name: [{ required: true, message: 'Name is required', trigger: 'blur' }],
+        role: [{ required: true, message: 'Role is required', trigger: 'change' }],
+        wecomId: [{ required: true, message: 'WeChat ID is required', trigger: 'blur' }],
+        status: [{ required: true, message: 'Status is required', trigger: 'change' }],
+        username: [
+          { required: true, message: 'Username is required', trigger: 'blur' },
+          { min: 4, message: 'Username must be at least 4 characters', trigger: 'blur' },
+        ],
+        password: [
+          { required: true, message: 'Password is required', trigger: 'blur' },
+          { min: 4, message: 'Password must be at least 4 characters', trigger: 'blur' },
+        ],
+        // newPassword: [
+        //   {
+        //     validator: (rule, value, callback) => {
+        //       console.log('Validating newPassword:', value);
+        //       if (this.changePassword && (!value || value.length < 4)) {
+        //         callback(new Error('Password must be at least 4 characters'));
+        //       } else {
+        //         callback();
+        //       }
+        //     },
+        //     trigger: 'blur',
+        //   },
+        // ],
+        // confirmPassword: [
+        //   {
+        //     validator: (rule, value, callback) => {
+        //       console.log('Validating confirmPassword:', value, 'Against:', this.newPassword);
+        //       if (this.changePassword && value !== this.newPassword) {
+        //         callback(new Error('Passwords must match'));
+        //       } else {
+        //         callback();
+        //       }
+        //     },
+        //     trigger: 'blur',
+        //   },
+        // ],
+      }
     };
   },
   created() {
@@ -293,13 +350,47 @@ export default {
       }
     },
     async handleStatusChange(userId, newStatus) {
+      const currentUserId = this.$store.getters.getUser.id;
+
+      // Check if deactivating self
+      if (userId === currentUserId && newStatus === 0) {
+        try {
+          await this.$confirm(
+              `You are deactivating your own account. Are you sure you want to proceed?`,
+              "Self Deactivation Confirmation",
+              {
+                confirmButtonText: "Yes",
+                cancelButtonText: "No",
+                type: "warning",
+              }
+          )
+              .then(async () => {
+                // Proceed with deactivation
+                const payload = { status: newStatus };
+                await updateUser(userId, payload);
+                this.$message.success("Your account has been deactivated successfully");
+                // Handle logout or session cleanup here
+              })
+              .catch(() => {
+                this.$message.info("Self-deactivation canceled");
+                this.fetchUserData(); // Refresh the table data
+              });
+        } catch (error) {
+          console.error("Error deactivating user:", error);
+          this.$message.error("Failed to deactivate your account");
+        }
+        return;
+      }
+
+      // Regular status change logic for other users
       try {
         const payload = { status: newStatus };
         await updateUser(userId, payload);
-        this.$message.success('Status updated successfully');
+        this.$message.success("Status updated successfully");
+        this.fetchUserData(); // Refresh the table data
       } catch (error) {
-        this.$message.error('Failed to update status');
-        console.error('Error updating status:', error);
+        console.error("Error updating status:", error);
+        this.$message.error("Failed to update status");
       }
     },
     getRoleName(roleId) {
@@ -311,7 +402,16 @@ export default {
       }
       return '未知角色';
     },
-    async handleAdd() {
+    validateAndAddUser() {
+      this.$refs.addUserForm.validate(async (valid) => {
+        if (valid) {
+          await this.performAddUser(); // Call the actual method to add the user
+        } else {
+          this.$message.error('Please fix validation errors before proceeding');
+        }
+      });
+    },
+    async performAddUser() {
       try {
         const roleId = this.newUser.role === '管理员' ? 1 : 2;
         const encryptedPassword = btoa(this.newUser.password);
@@ -328,40 +428,49 @@ export default {
 
         await addUser(payload);
         this.addDialogVisible = false;
-        this.fetchUserData();
+        await this.fetchUserData();
+        this.$message.success('User added successfully');
       } catch (error) {
         console.error('Error adding user:', error);
         this.$message.error('Failed to add user');
       }
     },
     async handleEditConfirm() {
-      try {
-        const roleId = this.editUser.role === '管理员' ? 1 : 2;
-        const payload = {
-          name: this.editUser.name,
-          role_id: roleId,
-          wecom_id: this.editUser.wecomId,
-          username: this.editUser.username,
-          email: this.editUser.email,
-          phone_number: this.editUser.phone_number,
-          status: this.editUser.status,
-        };
+      this.$refs.editUserForm.validate(async (valid) => {
+        if (valid) {
+          try {
+            const roleId = this.editUser.role === '管理员' ? 1 : 2;
+            const payload = {
+              name: this.editUser.name,
+              role_id: roleId,
+              wecom_id: this.editUser.wecomId,
+              username: this.editUser.username,
+              email: this.editUser.email,
+              phone_number: this.editUser.phone_number,
+              status: this.editUser.status,
+            };
 
-        // Include password if changePassword is checked
-        if (this.changePassword) {
-          if (!this.newPassword || this.newPassword !== this.confirmPassword) {
-            this.$message.error('Passwords do not match or are empty!');
-            return;
+            // Include password if changePassword is checked, in the future integrate to the same validation check
+            if (this.changePassword) {
+              if (!this.newPassword || this.newPassword !== this.confirmPassword) {
+                this.$message.error('Passwords do not match or are empty!');
+                return;
+              }
+              payload.password = btoa(this.newPassword);
+            }
+
+            await updateUser(this.editUser.id, payload);
+            this.editDialogVisible = false;
+            this.fetchUserData();
+            this.$message.success("User updated successfully");
+          } catch (error) {
+            console.error("Error editing user:", error);
+            this.$message.error("Failed to update user");
           }
-          payload.password = btoa(this.newPassword);
+        } else {
+          this.$message.error("Please fix validation errors before proceeding");
         }
-
-        await updateUser(this.editUser.id, payload);
-        this.editDialogVisible = false;
-        this.fetchUserData();
-      } catch (error) {
-        console.error('Error editing user:', error);
-      }
+      });
     },
     handleEdit(index, row) {
       this.editUser = {
@@ -380,11 +489,62 @@ export default {
       this.editDialogVisible = true;
     },
     async handleDelete(index, row) {
+      const currentUserId = this.$store.getters.getUser.id; // Get logged-in user ID
+
+      // Check if the user is deleting/deactivating themselves
+      if (row.id === currentUserId) {
+        try {
+          // Show self-deletion confirmation dialog
+          await this.$confirm(
+              `You are deleting your own account. Are you sure you want to proceed?`,
+              "Self Deletion Confirmation",
+              {
+                confirmButtonText: "Yes",
+                cancelButtonText: "No",
+                type: "warning",
+              }
+          )
+              .then(async () => {
+                // If confirmed, call delete API
+                await deleteUser(row.id);
+                this.$message.success("Your account has been deleted successfully");
+                this.fetchUserData(); // Refresh the table data
+                // handle logout or session cleanup: optional for now, give the user chance to wrap up
+              })
+              .catch(() => {
+                this.$message.info("Self-deletion canceled");
+                this.fetchUserData(); // Refresh the table data
+              });
+        } catch (error) {
+          console.error("Error deleting user:", error);
+          this.$message.error("Failed to delete your account");
+        }
+        return;
+      }
+
+      // Regular deletion for other users
       try {
-        await deleteUser(row.id);
-        this.fetchUserData(); // Refresh the data after deletion
+        this.$confirm(
+            `Are you sure you want to delete the user "${row.name}"?`,
+            "Delete Confirmation",
+            {
+              confirmButtonText: "Yes",
+              cancelButtonText: "No",
+              type: "warning",
+            }
+        )
+            .then(async () => {
+              // If confirmed, call delete API
+              await deleteUser(row.id);
+              this.$message.success("User deleted successfully");
+              this.fetchUserData(); // Refresh the table data
+            })
+            .catch(() => {
+              this.$message.info("Deletion canceled");
+            });
       } catch (error) {
-        console.error('Error deleting user:', error);
+        console.error("Error deleting user:", error);
+        this.$message.error("Failed to delete user");
       }
     },
     showAddDialog() {
@@ -413,4 +573,21 @@ export default {
   .popup-container {
     padding-right: 40px;
   }
+
+  .refresh-button {
+    background-color: #80cfff; /* Slightly lighter shade of primary color */
+    border-color: #80cfff; /* Match lighter background */
+  }
+
+  .refresh-button:hover {
+    background-color: #66b5ff; /* Slightly darker hover effect */
+    border-color: #66b5ff;
+    transform: rotate(360deg); /* Rotate on hover */
+    transition: transform 0.3s ease-in-out, background-color 0.2s ease; /* Smooth animation */
+  }
+
+  .refresh-button el-icon {
+    color: #004085; /* Darker primary-like color for the refresh icon */
+  }
+
 </style>
